change structure
put old download.py and update.py into utils.py
