+++ app/version.py
@@ -1,1 +1,1 @@
-__version__ = '1.8.8.14'
+__version__ = '1.8.8.1'
